+++ processor/src/test/resources/input/com/example/prop/NonnullChildPropComponent.java
@@ -1,0 +1,25 @@
+package com.example.prop;
+
+import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
+import react4j.annotations.Prop;
+import react4j.annotations.ReactComponent;
+import react4j.core.BaseState;
+import react4j.core.Component;
+import react4j.core.ReactNode;
+
+@ReactComponent
+abstract class NonnullChildPropComponent
+  extends Component<BaseState>
+{
+  @Nonnull
+  @Prop
+  protected abstract ReactNode getChild();
+
+  @Nullable
+  @Override
+  protected ReactNode render()
+  {
+    return null;
+  }
+}